Moved services form the services package to service Delete spv.mv.db
